requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
piexif>=1.1.3
exifread>=3.0.0
